+++ client/src/assets/style/NavBar.scss
@@ -1,0 +1,36 @@
+.nav-bar {
+	position: fixed;
+	bottom: 0;
+	left: 50%;
+	transform: translateX(-50%);
+	width: 390px;
+	height: 60px;
+	background-color: white;
+	display: flex;
+	justify-content: space-around;
+	align-items: center;
+	border-top: 1px solid #e0e0e0;
+	z-index: 1000;
+	box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.05);
+
+	button {
+		background: none;
+		border: none;
+		outline: none;
+
+		&.active svg {
+			color: #6fcf97;
+		}
+	}
+}
+
+.plus-button {
+	position: relative;
+	top: -15px;
+	background-color: #6fcf97;
+	border-radius: 50%;
+	padding: 8px;
+	svg {
+		color: #000000;
+	}
+}
