+++ client/src/assets/style/global.scss
@@ -22,3 +22,8 @@ body {
 	background: #ffffff;
 	box-shadow: 0 0 10px rgba(0, 0, 0, 0.05);
 }
+
+.content {
+	flex: 1;
+	padding-bottom: 70px;
+}
